% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{boot_mean_rel_interval}
\alias{boot_mean_rel_interval}
\title{Calcula um intervalo de confiança Bootstrap para diferença relativa de médias}
\usage{
boot_mean_rel_interval(tratamento, controle, B_size, replicas, inf, sup)
}
\arguments{
\item{tratamento}{Um vetor \code{numeric} com o quantidade estudada observada
no tratamento}

\item{controle}{Um vetor \code{numeric} com a quantidade estudada observada
no grupo controle}

\item{B_size}{Um número \code{integer} com o tamanho desejado para a reamostragem
Boostrap}

\item{replicas}{Um número \code{integer} com a quantidade de replicas (simulações)
Bootstrap que serão feitas}

\item{inf}{Um número \code{numeric} informando o quantil inferior do intervalo}

\item{sup}{Um número \code{numeric} informando o quantil inferior do intervalo}
}
\description{
Utilizando dois vetores \code{numeric} calcula-se o intervalo de
Bootstrap de diferença relativa \eqn{ \frac{mean(tratamento)}{mean(controle)}} de médias, conforme quantis fornecidos.
}
