% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_contagem_niveis.R
\name{misc_contagem_niveis}
\alias{misc_contagem_niveis}
\title{Contagem de eventos por um ID}
\usage{
misc_contagem_niveis(dados, var_id, var_temporal = NULL,
  var_categorica_alvo, subset_niveis_cat_alvo = NULL)
}
\arguments{
\item{dados}{Um objeto \code{data.table} que tenha pelo menos uma coluna identificando o ID do indivíduo e uma coluna com o evento}

\item{var_id}{Tipo \code{string} com o nome da coluna que identifica o indivíduo, um ID}

\item{var_temporal}{Tipo \code{string} com o nome da coluna que representa o ponto no tempo em que o evento ocorreu. DEFAULT: NULL}

\item{var_categorica_alvo}{Tipo \code{string} como o nome da coluna alvo, que possui os eventos}

\item{subset_niveis_cat_alvo}{Tipo vetor de com o tipo de sua variável. Caso queira um subset de niveis da categoria alvo os inform aqui}
}
\value{
Um objeto \code{data.table} Com a contagem de cada niveo para cada ID num formato WIDE.
}
\description{
Dado uma variável de evento, um ID e, opcionalmente, uma variável tempora. Calcula-se quantas vezes um evento ocorrou para determinado ID.
Caso tenha uma variável tempora será contado quantas unidades de tempo tiveram o evento do tipo (se houver dois eventos numa mesma unidade de tempo será contado como apenas um)
}
