% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_date.R
\name{misc_date}
\alias{misc_date}
\title{Identifica tipo de data}
\usage{
misc_date(dados, coluna_data, conferir = T)
}
\arguments{
\item{dados}{Um objeto \code{data.table} que tenha alguma coluna de strings de datas}

\item{conferir}{Um \code{logical} para conferir se há dois tipos diferentes de datas na mesma coluna. Se sim a função para. DEFAULT: T}

\item{colun_data}{Uma string com o nome da coluna em que está armazenada a data}
}
\value{
Nada. Altera INPLACE a coluna do \code{data.table} especificada
}
\description{
Identifica e transforma uma string para lubridate data (que use '/' ou '-') INPLACE nos formatos
YYYY-mm-dd, YYYY-dd-mm, YYYY-m-d, YYYY-d-m, mm-dd-YYYY, m-d-YYYY, dd-mm-YYYY, d-m-YYYY
YYYY/mm/dd, YYYY/dd/mm, YYYY/m/d, YYYY/d/m, mm/dd/YYYY, m/d/YYYY, dd/mm/YYYY, d/m/YYYY. E combinações entre mm, m, dd, d.
}
