% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_left_join.R
\name{misc_left_join}
\alias{misc_left_join}
\title{Left Join INPLACE com data.table}
\usage{
misc_left_join(dados_left, dados_right, columns_to_pull = NULL,
  columns_not_to_pull = NULL, by)
}
\arguments{
\item{dados_left}{Um objeto \code{data.table} que será o lado esquerdo da operação e será alterado INPLACE}

\item{dados_right}{Um objeto \code{data.table} que será o lado direito da operação, ele se manterá da mesma forma (com um index a mais)}

\item{columns_to_pull}{Um vetor de \code{string} contendo o nome das colunas que vc gostaria de puxar de dados_right.
Caso não seja especificado, será puxado todas.}

\item{columns_not_to_pull}{Um veto de \code{string} contendo o nome de colunas que vc não gostaria de puxar}

\item{by}{Um vetor nomeado indicando quais as colunas ID que será feito o Left Join. \code{by = c('lhs' = 'rhs')}}
}
\value{
Nada. A operação resultará na alteração INPLACE de \code{dados_left}
}
\description{
Realiza um Left Join INPLACE entre dois data.table por uma ou mais chaves
}
