% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msic_one_hot.R
\name{misc_one_hot}
\alias{misc_one_hot}
\title{Onr hot com data.table}
\usage{
misc_one_hot(dados, var_id, var_categorica_alvo)
}
\arguments{
\item{dados}{Um objeto \code{data.table}}

\item{var_id}{Uma \code{string} contendo o nome da variável identificadora}

\item{var_categorica_alvo}{Uma \code{string} com o nome da variável alvo. A variável que vc quer ver o one_hot}
}
\value{
Um objeto \code{data.table} com o Id e colunas para cada nível da categoria alvo
}
\description{
O objetivo desta função é, por meio de um objeto data.table com uma coluna identificadora do indivíduo
e uma variável categórica de interesse, criar um one_hot encoding.
}
