% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgb_treino_ensemble.R
\name{rnorm_t}
\alias{rnorm_t}
\title{Rnorm limitada}
\usage{
rnorm_t(n, mu, sd, upper = Inf, lower = -Inf)
}
\arguments{
\item{n}{O tamanho da amostra}

\item{mu}{A média da distribuição}

\item{sd}{O desvio padrão da distribuição}

\item{upper}{O limite superior da distribuição}

\item{lower}{O limite inferior da distribuição}
}
\value{
Retorna um sorteio aleatório de uma distribuição norma limitada nos valores superiores e inferiores dados para a função
}
\description{
Seu objetivo é gerar uma vizinhança, em torno dos parâmetros
}
