% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgb_predict_ensemble.R
\name{xgb_predict_ensemble}
\alias{xgb_predict_ensemble}
\title{xgb_predict_ensemble}
\usage{
xgb_predict_ensemble(folder, newdata, classe = F)
}
\arguments{
\item{folder}{Pasta onde estão armazenados os \code{n} modelos retornados função \code{xgb_treino_ensemble}}

\item{newdata}{Um objeto \code{data.table} com os novos dados para aplicação}

\item{classe}{Um indicador do que é retornado pelo modelo. Se uma classe ou um número contínuo: DEFALUT: F}
}
\value{
Retonar um objeto \code{data.table} com uma coluna nomeada \code{predicao}
}
\description{
Esta função tem como objetivo carregar os \code{n} modelos criados pela função \code{xgb_treino_ensemble}, aplicá-los aos dados criando a distribuição de previsão.
Após isso, caso seja regressão ou o retorno de uma probabilidade de classe, a previsão será a mediana, caso contrário a moda.
}
