% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgb_select_params.R
\name{xgb_select_params}
\alias{xgb_select_params}
\title{Selecionador de hyperparametros para Xgboost}
\usage{
xgb_select_params(dados, n_samples, var_id = NULL, metrica = "logloss",
  objetivo = "binary:logistic", niter_data = 10, niter_bayes = 50,
  cv.nfolds = 5, cv.nrounds = 3000, nthreads = 3,
  tree_method = "hist")
}
\arguments{
\item{dados}{Um objeto data.table com os dados pre processados e uma coluna target}

\item{n_samples}{O tamanho do sorteio par acada classe. Se regressão o tamanbho total da amostra}

\item{var_id}{A variável que identifica cada observação. Ela será removida antes da seleção de parametros. DEFAULT: \code{NULL} (não irá remover nada)}

\item{metrica}{A métrica selecionada para o xgboost lidar. DEFAULT: logloss}

\item{objetivo}{A finalidade do modelo. DEFAULT: binary:logistic}

\item{niter_data}{Quantas sessões serão feitas. Quantos sorteios serão feitos nos dados. DEFAULT: 10}

\item{niter_bayes}{Quantas itereções serão utilizadas no otimizador de bayes. DEFAULT: 50}

\item{cv.nfolds}{Quantos folds para o CV. DEFAULT: 5}

\item{cv.nrounds}{Quantidade máxima de nrounds a ser testado. DEFAULT: 3000,}

\item{nthreads}{Quantidade de CPU threads disponíveis para o XGBoost. DEFAULT: 3}

\item{tree_method}{Método de treino das árvores. DEFAULT: hist. Em hist será treinado na CPU, em gpu_hist na GPU}
}
\value{
Uma lista com três posições: Os parâmetros procurados, o número de nrounds e a melhor métrica encontrada
}
\description{
Seu objetivo é selecionar hyperparametros que maximizem ou minimizem seu objetivo utilizando CV e Sessões.
Para isto ele se utiliza do pacote mlrMBO, realizando optimizações de estatística bayesiana
}
