% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgb_treino_ensemble.R
\name{xgb_treino_ensemble}
\alias{xgb_treino_ensemble}
\title{Treino ensemble de modelos XGBoost}
\usage{
xgb_treino_ensemble(dados, n_samples, nthreads = 3, parametros = NULL,
  n_models = 100, save_importance = F, folder_to_save = NULL)
}
\arguments{
\item{dados}{Objeto \code{data.table} com os dados já pre processados e com uma coluna \code{target}.}

\item{n_samples}{O tamanho do sorteio par acada classe. Se regressão o tamanbho total da amostra}

\item{nthreads}{A quantidade de CPU threads disponíveis par ao XGBoost. DEFAULT: 3}

\item{parametros}{A lista de parâmetros retornado pela função \code{xgb_select_params}}

\item{n_models}{A quantidade de modelos gerados. Também é o número de sessões realizado. DEFAULT: 100}

\item{save_importance}{Indicador para o salvamento das importâncias das variáveis para cada modelo. DEFAULT: F}

\item{folder_to_save}{Pasta em que será salvo os \code{n_models} gerados}
}
\value{
A função não retorna nada, apenas salva os modelos na pasta
}
\description{
O objetivo desta função é treinar vários modelos por meio de sessões. Cada sessão possui uma parte dos dados, um modelo ali então é treinado e salvo numa pasta especificada.
}
